# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dispatch API
  version: 1.2.0
  description: Self-hosted notification orchestration, managed connector, and universal ingress API. API-key security applies when CONSOLE_AUTH_ENABLED=true; trusted-local mode is the development default.
servers:
  - url: http://localhost:8090
security:
  - apiKey: []
paths:
  /connect/v1/oauth/{connector}/callback:
    get:
      security: []
      summary: Consume a provider OAuth callback and redirect to the Integrations UI
      parameters:
        - {name: connector, in: path, required: true, schema: {type: string}}
        - {name: code, in: query, schema: {type: string}}
        - {name: state, in: query, schema: {type: string}}
        - {name: error, in: query, schema: {type: string}}
      responses:
        "303": {description: OAuth result redirect}
  /connect/v1/github/setup:
    get:
      security: []
      summary: Complete a state-protected GitHub App installation
      parameters:
        - {name: installation_id, in: query, required: true, schema: {type: integer, format: int64}}
        - {name: setup_action, in: query, schema: {type: string}}
        - {name: state, in: query, required: true, schema: {type: string}}
      responses:
        "303": {description: GitHub installation result redirect}
  /connect/v1/github/events:
    post:
      security: []
      summary: Receive an HMAC-authenticated GitHub App webhook delivery
      parameters:
        - {name: X-GitHub-Event, in: header, required: true, schema: {type: string}}
        - {name: X-GitHub-Delivery, in: header, required: true, schema: {type: string}}
        - {name: X-Hub-Signature-256, in: header, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json: {schema: {type: object, additionalProperties: true}}
      responses:
        "202": {description: Delivery accepted and routed}
        "401": {$ref: "#/components/responses/Error"}
...
  /api/v1/connectors/{connector}/authorize:
    post:
      summary: Begin a state- and PKCE-protected connector OAuth flow
      parameters:
        - {name: connector, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, recipient_id]
              properties:
                name: {type: string}
                recipient_id: {type: string}
                config:
                  type: object
                  additionalProperties: {type: string}
      responses:
        "200": {description: Provider authorization URL}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/connectors/github/install:
    post:
      summary: Begin a state-protected GitHub App installation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, recipient_id, config]
              properties:
                name: {type: string}
                recipient_id: {type: string}
                config:
                  type: object
                  required: [mode]
                  properties:
                    mode: {type: string, enum: [important, code, work, ci, all]}
      responses:
        "200": {description: GitHub App installation URL}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/connectors/telegram/auth/start:
    post:
      summary: Start Telegram user-account authorization and send a sign-in code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, recipient_id, phone_number]
              properties:
                name: {type: string}
                recipient_id: {type: string}
                phone_number: {type: string, pattern: '^\+[1-9][0-9]{6,14}$'}
      responses:
        "200": {description: Sign-in code requested}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "502": {$ref: "#/components/responses/Error"}
  /api/v1/connectors/telegram/auth/{authID}/code:
    post:
      summary: Verify a Telegram sign-in code
      parameters:
        - {name: authID, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties: {code: {type: string}}
      responses:
        "200": {description: 2FA password required}
        "201": {description: Telegram connection created}
        "404": {$ref: "#/components/responses/Error"}
        "502": {$ref: "#/components/responses/Error"}
  /api/v1/connectors/telegram/auth/{authID}/password:
    post:
      summary: Verify Telegram two-step authentication and create the connection
      parameters:
        - {name: authID, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties: {password: {type: string, format: password, writeOnly: true}}
      responses:
        "201": {description: Telegram connection created}
        "404": {$ref: "#/components/responses/Error"}
        "502": {$ref: "#/components/responses/Error"}
  /api/v1/connections:
    post:
      summary: Verify credentials and create a managed connector connection
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateConnectorConnection"}
      responses:
        "201": {description: Connection created; credentials are not returned}
        "400": {$ref: "#/components/responses/Error"}
        "502": {$ref: "#/components/responses/Error"}
  /api/v1/connections/{id}/test:
    post:
      summary: Recheck credentials and retry connector activation
      parameters: [{$ref: "#/components/parameters/ID"}]
      responses:
        "200": {description: Test result and activation state}
        "502": {$ref: "#/components/responses/Error"}
  /api/v1/connections/{id}/sample:
    post:
      summary: Send a sample event through the normal Dispatch delivery pipeline
      parameters: [{$ref: "#/components/parameters/ID"}]
      responses:
        "202": {description: Sample notification queued}
  /api/v1/connections/{id}/enabled:
    post:
      summary: Activate or pause a connector and its external subscription
      parameters: [{$ref: "#/components/parameters/ID"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties: {enabled: {type: boolean}}
      responses:
        "204": {description: Lifecycle state updated}
        "502": {$ref: "#/components/responses/Error"}
  /api/v1/connections/{id}:
    put:
      summary: Update a connection name, recipient, and user-editable preferences
      parameters: [{$ref: "#/components/parameters/ID"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateConnectorConnection"}
      responses:
        "200": {description: Connection updated}
        "409": {$ref: "#/components/responses/Error"}
    delete:
      summary: Deactivate and delete a connector connection
      parameters: [{$ref: "#/components/parameters/ID"}]
      responses:
        "204": {description: Disconnected}
        "502": {$ref: "#/components/responses/Error"}
  /api/v1/connections/{id}/authorize:
    post:
      summary: Reauthorize an OAuth connection and update it in place
      parameters: [{$ref: "#/components/parameters/ID"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateConnectorConnection"}
      responses:
        "200": {description: Authorization URL}
        "409": {$ref: "#/components/responses/Error"}
  /ingest/v1/{slug}:
    post:
      security: []
      summary: Authenticate, transform, and idempotently ingest a source event
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: true}
      responses:
        "202": {description: Notification created and queued}
        "200": {description: Duplicate event or provider verification response}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/sources:
    get:
      summary: List universal ingress sources and available presets
      responses:
        "200":
          description: Sources
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {type: array, items: {$ref: "#/components/schemas/IngressSource"}}
                  providers: {type: array, items: {type: string}}
    post:
      summary: Create a source and return its secret once
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateIngressSource"}
      responses:
        "201": {description: Source created; response includes the one-time secret}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/sources/{id}/rotate-secret:
    post:
      summary: Rotate a source secret and return the new value once
      parameters: [{$ref: "#/components/parameters/ID"}]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                secret: {type: string, writeOnly: true, description: Required for Slack and Stripe}
      responses:
        "200": {description: Secret rotated}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/sources/{id}/enabled:
    post:
      summary: Enable or disable a source
      parameters: [{$ref: "#/components/parameters/ID"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties: {enabled: {type: boolean}}
      responses:
        "204": {description: Updated}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/sources/{id}/events:
    get:
      summary: List recent accepted events for a source
      parameters: [{$ref: "#/components/parameters/ID"}]
      responses:
        "200": {description: Source event history}
...
components:
  securitySchemes:
    apiKey: {type: apiKey, in: header, name: X-API-Key}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string}}
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                required: [code, message]
                properties: {code: {type: string}, message: {type: string}}
  schemas:
...
    CreateConnectorConnection:
      type: object
      required: [connector_id, name, recipient_id]
      properties:
        connector_id: {type: string}
        name: {type: string}
        recipient_id: {type: string}
        config: {type: object, additionalProperties: {type: string}}
        credentials:
          type: object
          writeOnly: true
          additionalProperties: {type: string}
    UpdateConnectorConnection:
      type: object
      required: [name, recipient_id, config]
      properties:
        name: {type: string}
        recipient_id: {type: string}
        config: {type: object, additionalProperties: {type: string}}
...